pyproject.toml
[tool.poetry]
name = "dllab-ddd-sample"
version = "0.1.0"
description = ""
authors = ["wf-yamaday <e155417k@st.u-gakugei.ac.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
django-ninja = "^0.19.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-django = "^4.5.2"
pysen = {version = "0.10.1"}
mypy = "^0.931"
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
django-stubs = "^1.9.0"

[tool.pysen]
version = "0.10.1"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
mypy_path = ["./sample_project"]
line_length = 88
py_version = "py38"

[tool.pysen.lint.source]
  includes = ["./sample_project"]
  exclude_globs = ["**/migrations/*.py"]

[[tool.pysen.lint.mypy_targets]]
  paths = ["./sample_project"]

[[tool.pysen.lint.mypy_plugins]]
  function = "mypy_django_plugin.main"
[[tool.pysen.lint.mypy_plugins]]
  function = "pydantic.mypy"

[tool.pysen-cli]
settings_dir = "."

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "-s -v -ra --durations=0"
python_files =["tests.py","test_*.py","*_tests.py"]
testpaths = [
    "sample_project",
]

[tool.black] # automatically generated by pysen
line-length = 88
target-version = ["py38"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
